refactor: Move seed to Scrap folder and remove ! to event class
