+++ src/bot/classes/events.ts
@@ -3,8 +3,8 @@ import { ClientEvents } from "discord.js";
 import CustomClient from "./client";
 
 export class Event {
-    name!: keyof ClientEvents;
-    run!: (client: CustomClient, ...eventArgs: any) => void;
+    name: keyof ClientEvents;
+    run: (client: CustomClient, ...eventArgs: any) => void;
     constructor(options: NonNullable<Event>) {
         Object.assign(this, options);
     }
