+++ src/scrap/scrapWorker.ts
@@ -1,19 +1,19 @@
-import { parentPort, threadId } from 'worker_threads';
-import { Browser, Page } from 'puppeteer';
-import { startBrowser } from './browser';
-import { MangaInfo, ScrapingResult, ScrapingError } from '../types/types';
-import { getChapterElement } from '../API/seed';
+import { parentPort, threadId } from "worker_threads";
+import { Browser, Page } from "puppeteer";
+import { startBrowser } from "./browser";
+import { MangaInfo, ScrapingResult, ScrapingError } from "../types/types";
+import { getChapterElement } from "./seed";
 
 (async () => {
     const browser: Browser = await startBrowser();
 
-    parentPort?.on('message', async (task: { manga: MangaInfo }) => {
+    parentPort?.on("message", async (task: { manga: MangaInfo }) => {
         const { manga } = task;
         const page: Page = await browser.newPage();
 
         try {
             let maxChapter = Number(manga.chapter);
-            let maxChapterSite: MangaInfo['sites'][number] | null = null;
+            let maxChapterSite: MangaInfo["sites"][number] | null = null;
             let maxChapterURL = "";
             let lastChapterText = "";
 
@@ -25,15 +25,15 @@ import { getChapterElement } from '../API/seed';
                 lastChapterText = await getChapterElement(page, site.chapter_url.split("/").at(-2) ?? "", site, manga);
 
                 const lastChapterTextMatch = lastChapterText
-                    ?.replace(/\/$/, '')
-                    ?.split('/')
+                    ?.replace(/\/$/, "")
+                    ?.split("/")
                     .at(-1)
                     ?.match(/(\d+(?:[\.-]\d+)?)/);
-                const lastChapter = lastChapterTextMatch
-                    ? parseFloat(lastChapterTextMatch[0].replace('-', '.'))
-                    : NaN;
+                const lastChapter = lastChapterTextMatch ? parseFloat(lastChapterTextMatch[0].replace("-", ".")) : NaN;
 
-                console.log(`[${new Date().toLocaleString()}] Worker ${threadId} Scraped ${manga.name} at ${site.url}: ${lastChapter}`);
+                console.log(
+                    `[${new Date().toLocaleString()}] Worker ${threadId} Scraped ${manga.name} at ${site.url}: ${lastChapter}`
+                );
 
                 if (!isNaN(lastChapter)) {
                     if (lastChapter > maxChapter) {
@@ -46,7 +46,7 @@ import { getChapterElement } from '../API/seed';
 
             if (maxChapterSite !== null) {
                 parentPort?.postMessage({
-                    type: 'result',
+                    type: "result",
                     data: {
                         manga,
                         lastChapter: maxChapter.toString(),
@@ -56,13 +56,13 @@ import { getChapterElement } from '../API/seed';
                 });
             } else {
                 parentPort?.postMessage({
-                    type: 'error',
-                    data: { name: manga.name, error: 'Failed to scrape any site for updates.' } as ScrapingError,
+                    type: "error",
+                    data: { name: manga.name, error: "Failed to scrape any site for updates." } as ScrapingError,
                 });
             }
         } catch (error: any) {
             parentPort?.postMessage({
-                type: 'error',
+                type: "error",
                 data: { name: manga.name, error: error.message } as ScrapingError,
             });
         } finally {
@@ -70,7 +70,7 @@ import { getChapterElement } from '../API/seed';
         }
     });
 
-    process.on('exit', async () => {
+    process.on("exit", async () => {
         await browser.close();
     });
 })();
